% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{addGuards}
\alias{addGuards}
\title{Add guards to column names}
\usage{
addGuards(x)
}
\arguments{
\item{x}{- character vector to convert}
}
\value{
character vector with guards added (as necessary)
}
\description{
Function to add guards to column names
}
