% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byGridCell_AddValues.R
\name{byGridCell_AddValues}
\alias{byGridCell_AddValues}
\title{Add values by grid cell}
\usage{
byGridCell_AddValues(
  dfr1,
  dfr2,
  cols = c("unsuccessful_indivs", "successful_indivs", "total_indivs",
    "unsuccessful_abundance", "successful_abundance", "total_abundance")
)
}
\arguments{
\item{dfr1}{- dataframe with results by grid cell}

\item{dfr2}{- dataframe with results by grid cell}

\item{cols}{- dataframe column names to add together}
}
\value{
dataframe (dfr1) with added results in requested columns.
}
\description{
Function to add values by grid cell.
}
\details{
"Factor" levels (gridCellID, startTime) in dataframes must be aligned,
each must have the same number of rows.
}
