% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byGridCell_CreateSFDataset.R
\name{byGridCell_CreateSFDataset}
\alias{byGridCell_CreateSFDataset}
\title{Create a map layer with values by grid cell}
\usage{
byGridCell_CreateSFDataset(dfr, roms_grid)
}
\arguments{
\item{dfr}{- dataframe with values to create map layer by gridCellID}

\item{roms_grid}{- shapefile name or layer (sf dataset) with grid polygons by ID}
}
\value{
an \code{sf} (simple features) dataframe with the values by grid cell, or NULL if all values are 0.
}
\description{
Function to create a map layer with values by grid cell.
}
\details{
Uses package \code{wtsGIS}. If all values are 0, then NULL is
returned for the sf dataframe object.
}
