% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byGridCell_GridOpsOnGrids.R
\name{byGridCell_GridOpsOnGrids}
\alias{byGridCell_GridOpsOnGrids}
\title{Perform an algebraic operation on values by grid cell in two dataframes}
\usage{
byGridCell_GridOpsOnGrids(
  dfr1,
  op,
  dfr2,
  cols = c("unsuccessful_indivs", "successful_indivs", "total_indivs",
    "unsuccessful_abundance", "successful_abundance", "total_abundance")
)
}
\arguments{
\item{dfr1}{- dataframe with results by grid cell}

\item{op}{- operation by grid cell ("+","-","*","/")}

\item{dfr2}{- dataframe with results by grid cell}

\item{cols}{- dataframe column names to add together}
}
\value{
dataframe (dfr1) with results of operation in requested columns.
}
\description{
Function to perform an algebraic operation on values by grid cell in two dataframes.
}
\details{
"Factor" levels (gridCellID, startTime) in dataframes must be aligned,
each must have the same number of rows.#' Columns in datafram not specified in
\code{cols} are returned unchanged.
}
