% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byGridCell_ScalarOps.R
\name{byGridCell_ScalarOps}
\alias{byGridCell_ScalarOps}
\title{Perform a scalar operation on grid cell values}
\usage{
byGridCell_ScalarOps(
  dfr1,
  op = "+",
  scalar = 0,
  cols = c("unsuccessful_indivs", "successful_indivs", "total_indivs",
    "unsuccessful_abundance", "successful_abundance", "total_abundance")
)
}
\arguments{
\item{dfr1}{- dataframe with results by grid cell}

\item{op}{- operation to perform ("+","-","*","/","\","^","log")}

\item{scalar}{- scalar value to use}

\item{cols}{- dataframe column names to scale}
}
\value{
dataframe with results in requested columns.
}
\description{
Function to perform a scalar operation on grid cell values.
}
\details{
Columns in datafram not specified in \code{cols} are returned unchanged.
"/" performs grid/scalar whereas \" performs scalar/grid.
}
