% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byGridCell_ScaleValues.R
\name{byGridCell_ScaleValues}
\alias{byGridCell_ScaleValues}
\title{Scale values by grid cell by a multiplicative constant}
\usage{
byGridCell_ScaleValues(
  dfr1,
  scale,
  cols = c("unsuccessful_indivs", "successful_indivs", "total_indivs",
    "unsuccessful_abundance", "successful_abundance", "total_abundance")
)
}
\arguments{
\item{dfr1}{- dataframe with results by grid cell}

\item{scale}{- multiplicative scale factor}

\item{cols}{- dataframe column names to scale}
}
\value{
dataframe with scaled results
}
\description{
Function to scale values by grid cell by a multiplicative constant.
}
\details{
None.
}
