% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvgAbundanceByGridCell.R
\name{calcAvgAbundanceByGridCell}
\alias{calcAvgAbundanceByGridCell}
\title{Calculate average abundance of individuals by grid cell}
\usage{
calcAvgAbundanceByGridCell(dfrs, byStartTime = FALSE)
}
\arguments{
\item{dfrs}{- list of dataframes, by typeName, with DisMELS IBM results}

\item{byStartTime}{- flag to average by startTime}
}
\value{
a list of dataframes, by life stage, with average abundance by grid cell
}
\description{
Function to calculate average abundance of individuals by grid cell.
}
\details{
Uses packages \code{sqldf}, \code{reshape2}.
}
