% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConnectivityMatrix.R
\name{calcConnectivityMatrix}
\alias{calcConnectivityMatrix}
\title{Calculate a connectivity matrix dataframe from a connectivity results \pkg{sf} dataframe}
\usage{
calcConnectivityMatrix(
  sfs_points,
  sf_conn,
  startStage = "Z1",
  endStage = "C1F",
  endStageFac = 2,
  startZones = c(1:8),
  endZones = c(1:18),
  plotStartLocs = FALSE,
  plotEndLocs = FALSE,
  bmls = NULL,
  crs = wtsGIS::get_crs(4326),
  nSZpG = 2,
  nPlotCols = 2,
  colours = c("red", "blue", "green", "cyan", "black")
)
}
\arguments{
\item{sfs_points}{- list by life stage of \pkg{sf} dataframes with connectivity results and point locations from a single model run}

\item{sf_conn}{- \pkg{sf} dataframe with connectivity zone information (polygons)}

\item{startStage}{- name of start life stage for connectivity matrix}

\item{endStage}{- name of end life stage for connectivity matrix}

\item{endStageFac}{- multiplier on apparent number successful (e.g., 2 if sexes were split after starting life stage)}

\item{startZones}{- vector of starting zone id's (integers) for sf_conn}

\item{endZones}{- vector of ending zone id's (integers) for sf_conn}

\item{plotStartLocs}{- flag to plot start locations}

\item{plotEndLocs}{- flag to plot end locations}

\item{bmls}{- list of \pkg{ggplot2} basemap layers}

\item{crs}{- \pkg{sf} coordinate reference system (crs) for maps}

\item{nSZpG}{- number of start zones to include in a group for end location maps}

\item{nPlotCols}{- number of columns for end location maps}

\item{colours}{- vector of colours to use to distinguish start zones in an end location map}
}
\value{
list with elements
\itemize{
  \item{step3_StartEndZones - dataframe with start/end zones for each individual}
  \item{step3_ConnectivityDataframe - connectivity matrix as dataframe}
  \item{pStartMap - ggplot2 map of start locations (or NULL)}
  \item{pEndMap - ggplot2 map of end locations (or NULL)}
}
}
\description{
Function to calculate a connectivity matrix datafarme from a connectivity results dataframe.
}
\details{
Calculates the connectivity matrix for a model run based on the connectivity
zones defined in \code{sf_conn}. Optionally, \pkg{ggplot2}-style maps of the starting and
ending locations of individuals can be created. End zone locations are colored according
to their start zone.
}
