% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConnectivityMatrixAlt.R
\name{calcConnectivityMatrixAlt}
\alias{calcConnectivityMatrixAlt}
\title{Calculate a connectivity matrix dataframe classified by start/end zones}
\usage{
calcConnectivityMatrixAlt(
  dfr_isezs,
  endLHSs = c("C1F", "C1M"),
  startZones = c(1:8),
  endZones = c(1:18)
)
}
\arguments{
\item{dfr_isezs}{- dataframe with individuals classified by start/end zones}

\item{endLHSs}{- name(s) of ending life stage(s) for connectivity matrix}

\item{startZones}{- vector of all starting zone id's (integers)}

\item{endZones}{- vector of all ending zone id's (integers)}
}
\value{
connectivity matrix as a dataframe
}
\description{
Function to calculate a connectivity matrix dataframe individuals classified by start/end zones.
}
\details{
Calculates the connectivity matrix for a model run based on individuals previously
classified by start and end zones (e.g., from \code{\link{indivsInfo_ExtractStartEndByStage}}
or \code{\link{indivsInfo_ExtractStartEnd}}, followed by \code{\link{indivsInfo_AssignZones}}).

This function provides an alternative path to calculating the same connectivity matrices
as \code{\link{calcConnectivityMatrix}}.
}
