% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConnectivityMatrixDiffs.R
\name{calcConnectivityMatrixDiffs}
\alias{calcConnectivityMatrixDiffs}
\title{Calculate differences between connectivity matrices}
\usage{
calcConnectivityMatrixDiffs(
  tbl_conns,
  tbl_base,
  conns_col = "connFrac",
  base_col = "meanFrac"
)
}
\arguments{
\item{tbl_conns}{- dataframe with connectivity values}

\item{tbl_base}{- dataframe with base connectivity values}

\item{conns_col}{- name of column in tbl_conns to difference}

\item{base_col}{- name of column in tbl_base for differencing}
}
\value{
A dataframe of connectivity matrices idential to \code{tbl_conns}, but
with the requested difference values in an additional column, \code{diff}.
}
\description{
Function to calculate differences between connectivity matrices.
}
\details{
Connectivity values from the base matrix (iei., \code{tbl_base}) are
subtracted from corresponding values in \code{tbl_conns} on the basis of matching
values in the startZone and endZone columns in each dataframe.

Note that \code{tbl_conns} may encapsulate several connectivity matrices. The base
matrix will be subtracted from each.
}
