% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMeanConnectivityMatrix.R
\name{calcMeanConnectivityMatrix}
\alias{calcMeanConnectivityMatrix}
\title{Calculate mean connectivity matrix}
\usage{
calcMeanConnectivityMatrix(tbl_conn)
}
\arguments{
\item{tbl_conn}{- dataframe with connectivity information from which to extract means by startZone **x** endZone}
}
\value{
a dataframe with mean, median, and std dev. of connectivity values
}
\description{
Function to calculate a mean connectivity matrix (as a dataframe).
}
\details{
Mean, median, and std deviation of the \code{connFrac} column of the input
\code{tbl_conn} are calculated by startZone **x** endZone
}
