% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConnectivityZones.R
\name{checkConnectivityZones}
\alias{checkConnectivityZones}
\title{Check connectivity zones for possible problems}
\usage{
checkConnectivityZones(sf_zones)
}
\arguments{
\item{sf_zones}{- \pkg{sf} dataframe with connectivity zones (polygons)}
}
\value{
list with elements 'validity' and 'overlaps', containing output from running
\code{\link[sf]{st_is_valid}} and \code{\link[sf]{st_overlaps}}, respectively,
on \code{sf_zones}.
}
\description{
Function to check connectivity zones for possible problems.
}
\details{
Ideally, \code{sf_zones} should consist of non-overlapping polygons so that
only one zone can be assigned to each geometry. If a point geometry is in an area of overlap
between multiple zones, a record would normally be included in the result for each zone
in which the point falls. If \code{keepFirst} is TRUE, the record for the first zone
in which the point falls is kept. This is also the case if zones are assigned to multiple
\code{geomNames}. Otherwise, records are added to the result for all zones in which a point falls.
}
