% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyIndivsBySuccess.R
\name{classifyIndivsBySuccess}
\alias{classifyIndivsBySuccess}
\title{Extract IDs for successful individuals}
\usage{
classifyIndivsBySuccess(dfrStart, dfrEnd, typeName)
}
\arguments{
\item{typeName}{- name of life stage defining success}

\item{dfr}{- a dataframe or tibble with the "successful" life stage}
}
\value{
a dataframe or tibble with origID, startTime, and successful for all individuals.
}
\description{
Function to extract IDs for successful individuals.
}
\details{
The
}
