% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyIndivsBySuccess_OLD.R
\name{classifyIndivsBySuccess_OLD}
\alias{classifyIndivsBySuccess_OLD}
\title{Classify individuals by success/failure}
\usage{
classifyIndivsBySuccess_OLD(dfrs, index = NULL, typeName)
}
\arguments{
\item{dfrs}{- list of dataframes by typeName}

\item{index}{- dataframe returned from \code{extractIndexForSuccessfulIndivs}}

\item{typeName}{- life stage typeName considered successful (an alternative to providing \code{index})}
}
\value{
the list of dataframes by typeName, with added column "succcess"
(TRUE/FALSE) indicating whether individual is considered a success.
}
\description{
Function to classify individuals by success/failure.
}
\details{
Successful individuals are identified using \code{extractIndexForSuccessfulIndivs(dfrs,typeName)} if
\code{index} is null.
}
