% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConnectivityWeb.R
\name{createConnectivityWeb}
\alias{createConnectivityWeb}
\title{Create a \pkg{sf} dataframe with lines between connectivity zones}
\usage{
createConnectivityWeb(
  dfr,
  xCol = "lon",
  yCol = "lat",
  crs = sf::NA_crs_,
  wrapDateline = TRUE
)
}
\arguments{
\item{dfr}{- a dataframe (or \pkg{sf} dataframe) with lat/lon or xy locations representing connectivity zones}

\item{xCol}{- name of column with x-axis (longitude) location}

\item{yCol}{- name of column with y-axis (latitude) location}

\item{crs}{- coordinate reference system (object that can be converted to an \code{sf::crs} object; default is NA_crs_)}

\item{wrapDateline}{- flag (T/F) to use 0 to 360 rather than -180 to 180 range for longitudes}
}
\value{
a \pkg{sf} dataframe. The 'geometry' column contains lines which connect
point locations identifying connectivity zones.
}
\description{
Function to create a \pkg{sf} dataframe with lines between connectivity zones
}
\details{
Creates a \pkg{sf} dataframe with a "web" of directed lines connecting connectivity
zones. Coordinates for "start" and "end" zones are also given. If \code{dfr} is an \pkg{sf}
dataframe, then the input value for \code{crs} is ignored and taken from \code{dfr}.
}
