% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMapLayer_ValuesByGridCell.R
\name{createMapLayer_ValuesByGridCell}
\alias{createMapLayer_ValuesByGridCell}
\title{Create a map layer with values by grid cell}
\usage{
createMapLayer_ValuesByGridCell(
  dfr,
  vals_col,
  basemap = NULL,
  title = "values",
  alpha = 0.9,
  plotMap = FALSE
)
}
\arguments{
\item{dfr}{- \code{sf} dataframe with values to create map layer by gridCellID}

\item{vals_col}{- column name of values to plot}

\item{basemap}{- basemap to use (default is NULL)}

\item{title}{- map title (default is "values")}

\item{alpha}{- transparency to apply to values layer (default is 0.9)}

\item{plotMap}{- flag to combine layer with basemap and plot (default is FALSE)}
}
\value{
a tmap layer object of the values by grid cell.
}
\description{
Function to create a map layer with values by grid cell.
}
\details{
Requires package \code{tmap}. If all values are 0, then NULL is
returned for the map layer object. Note that the color scale used for the values
is determined by the \code{aes.palette} argument to \code{createBasemap} when the basemap is created.
}
