% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMapLayers_Locations.R
\name{createMapLayers_Locations}
\alias{createMapLayers_Locations}
\title{Create map layers from a list of dataframes with point locations}
\usage{
createMapLayers_Locations(
  dfrs_points,
  markers = c("dots", "symbols"),
  shape = 1,
  size = 0.2,
  alpha = 0.6,
  basemap = NULL,
  colors.background = "grey85",
  colors.land = "grey45",
  colors.bathym = "grey65",
  aes.palette = list(cat = c("blue", "red"), div = c("blue", "red"), seq = c("blue",
    "red")),
  printMaps = FALSE
)
}
\arguments{
\item{markers}{- marker type (e.g., "dots", "symmbols")}

\item{shape}{- integer id for shape type}

\item{size}{-marker size}

\item{alpha}{-marker transparency}

\item{basemap}{- basemap on which to plot locations (if printMaps is TRUE)}

\item{colors.background}{- basemap background color passed to \code{createBasemap}}

\item{colors.land}{-basemap land color passed to \code{createBasemap}}

\item{colors.bathym}{- basemap color for bathymetric contours, passed to \code{createBasemap}}

\item{aes.palette}{- basemap list for aes.palette passed to \code{createBasemap}}

\item{printMaps}{- flag to print maps (basemap + each layer)}

\item{dfr_points}{- dataframe with point locations}
}
\value{
list of map layers corresponding to different life stage types
}
\description{
Function to create map layers from a list of dataframes with point locations.
}
\details{
Requires package \code{tmap}. If basemap is NULL and printMaps is TRUE,
a basemap is created
using a call to \code{createBasemap} with the supplied parameters.
}
