% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSFDatasets_Points.R
\name{createSFDatasets_Points}
\alias{createSFDatasets_Points}
\title{Create a list of sf (simple features) datasets (dataframes) with point geometries from DisMELS tracks}
\usage{
createSFDatasets_Points(dfrs)
}
\arguments{
\item{dfrs}{- list of dataframes, by typeName, with DisMELS IBM results}

\item{strCRS}{- coordinate reference system: EPSG code or character with proj4string}
}
\value{
a list of sf dataset objects, by typeName, each with a column ("geom") of class sfc_POINT
}
\description{
Function to create a list of sf (simple features) datasets (dataframes) with point geometries from DisMELS tracks.
}
\details{
Uses packages \code{sf}, \code{wtsGIS}. For each sf dataset, the point geometries
are in column "geom". Other columns are copied from the associated input dataframes.
}
