% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSFDatasets_Trajectories.R
\name{createSFDatasets_Trajectories}
\alias{createSFDatasets_Trajectories}
\title{Create a list of sf (simple features) datasets (dataframes) with linestring geometries from DisMELS tracks}
\usage{
createSFDatasets_Trajectories(dfrs, strCRS = wtsGIS::getCRS("WGS84"))
}
\arguments{
\item{dfrs}{- list of dataframes, by typeName, with DisMELS IBM results}

\item{strCRS}{- coordinate reference system: EPSG code or character with proj4string}
}
\value{
a list of sf dataset objects, each with a column ("geom") of class sfc_LINESTRING
}
\description{
Function to create a list of sf (simple features) datasets (dataframes) with linestring geometries from DisMELS tracks.
}
\details{
Uses packages \code{sf}, \code{wtsGIS}. For each sf dataset, the linestring geometry
is in column "geom". Other columns include id, startTime, age, ageInStage, and successful.
}
