% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIndexForSuccessfulIndivs.R
\name{extractIndexForSuccessfulIndivs}
\alias{extractIndexForSuccessfulIndivs}
\title{Extract an index for successful individuals}
\usage{
extractIndexForSuccessfulIndivs(dfrs, typeName)
}
\arguments{
\item{dfrs}{- dataframe with successful individuals or list of dataframes}

\item{typeName}{- name of life stage defining success}
}
\value{
a dataframe with origID and startTime for successful individuals
}
\description{
Function to extract an index for successful individuals.
}
\details{
The "index" consists of a dataframe listing the origID and startTime for
each successful individual in dfrs (or in dfrs[[typeName]] if dfrs is a list of dataframes).
}
