% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_extractIndivs.R,
%   R/extractTrajectories.R
\name{extractTrajectories}
\alias{extractTrajectories}
\title{Extract trajectories of (selected) individuals from a DisMELS model run}
\usage{
extractTrajectories(dfrs, strCRS = wtsGIS::getCRS("WGS84"))

extractTrajectories(dfrs, strCRS = wtsGIS::getCRS("WGS84"))
}
\arguments{
\item{dfrs}{- list of dataframes, by typeName, with DisMELS IBM results}

\item{strCRS}{- coordinate reference system: EPSG code or character with proj4string}

\item{results}{- DisMELS results file (or dataframe)}

\item{indivIDs}{- vector of id's for individuals to extract (NULL = all)}

\item{lsTypesInfo}{- life stage types info list for IBM}

\item{lsTypes}{- vector of names of life stage types to extract (NULL = all)}

\item{returnList}{- flag (T/F) indicating whether a list should be returned (otherwise NULL--helps with memory)}

\item{writeOutput}{- flag (T/F) indicating whether output files should be written}

\item{outDir}{- folder for output}

\item{outBaseCSV}{- base name for output csv files}

\item{verbose}{- flag (T/F) to print debugging info}
}
\value{
list of dataframes with individual results by lhs type name if returnList=TRUE, otherwise NULL.

a list of sf dataset objects, each with a column ("geom") of class sfc_LINESTRING
}
\description{
Function to extract trajectories of (selected) individuals from a DisMELS model run.

Function to extract trajectories from DisMELS output as a list of sf (simple features) datasets (dataframes) with linestring geometries.
}
\details{
If results is NULL, the user will be prompted to select the
corresponding file using a file chooser dialog.
Uses functions \code{wtsUtilities::getCSV} and \code{sqldf::sqldf}.

Uses packages \code{sf}, \code{wtsGIS}. For each sf dataset, the linestring geometry
is in column "geom". Other columns include id, startTime, age, ageInStage, and successful.
}
