% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandardAttributes.R
\name{getStandardAttributes}
\alias{getStandardAttributes}
\title{Get standard attribute names for DisMELS output}
\usage{
getStandardAttributes(resType)
}
\arguments{
\item{resType}{- flag indicating DisMELS results type ("Latest","NEW2.0SC","NEW2.0", or "OLD")}
}
\value{
data frame with columns for standard attribute names ('name'), short names ('short_name'), and data types ('data_type')
}
\description{
Function to get standard attribute names for DisMELS output from several versions of DisMELS.
}
\details{
Use \code{resType=}\cr
 \code{"Latest"} for model results with latest DisMELS version\cr
 \code{"NEW2.0SC"} for model results from DisMELS2.0_SnowCrab\cr
 \code{"NEW2.0"}   for model results from DisMELS2.0\cr
 \code{"OLD"}      for model results from (original) DisMELS\cr
}
