% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMap_IndivTrajectories.R
\name{ggMap_IndivTrajectories}
\alias{ggMap_IndivTrajectories}
\title{Plot trajectories from a DisMELS model run by start zone}
\usage{
ggMap_IndivTrajectories(
  sf_trjs,
  colorBy = "successful",
  colorLabel = colorBy,
  subtitle = "",
  bmls,
  startZones = 1:8,
  nSZpG = 1,
  nPlotCols = 4,
  colours = c("red", "cyan", "green", "blue", "black")
)
}
\arguments{
\item{sf_trjs}{- \pkg{sf} dataframe from \code{\link{indivsInfo_ExtractTrajectories}}}

\item{colorBy}{- name of column to use as levels to color tracks by}

\item{colorLabel}{- label for colours legend}

\item{subtitle}{- plot subtitle}

\item{bmls}{- basemap layers object (e.g., from \code{\link[rIBMsSnowCrab]{getBasemapLayers}})}

\item{startZones}{- vector of valid start zones}

\item{nSZpG}{- number of start zones per map group (1 map is plotted per group)}

\item{nPlotCols}{- number of columns of maps per page}

\item{colours}{- vector of colours to use for trajectories}
}
\value{
ggplot2 object
}
\description{
Function to plot trajectories from a DisMELS model run by start zone.
}
\details{
Separate maps are created for each group of start zones (in column "startZone"),
then arranged using \code{\link[ggplot2]{facet_wrap}}. The total number of maps per page
will be the number of \code{startZones/nSZpG}, arranged into \code{nPlotCols} columns.

The colour values in \code{colours} will be truncated or repeated to match
the number of distinct levels in the column specified by \code{colorBy}. It is
probably best if this column is a factor, with the expectation that factor levels
map to colours by index into the truncated/expanded \code{colours}.
}
