% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMap_sf.R
\name{ggMap_sf}
\alias{ggMap_sf}
\title{TODO: finish. Create a ggplot2 object with maps from a \pkg{sf} dataframe}
\usage{
ggMap_sf(
  sf_dfr,
  geomsCol = "geometry",
  colorCol = NULL,
  colorScale = c("discrete", "continuous"),
  colours = c("red", "blue", "green", "cyan", "black"),
  fillCol = NULL,
  fillScale = c("discrete", "continuous"),
  fills = c("red", "blue", "green", "cyan", "black"),
  shapeCol = NULL,
  shapes = c(20, 21, 22),
  alpha = 0.5,
  facet_grid = NULL,
  facet_grid_scales = "fixed",
  facet_wrap = NULL,
  bmls = NULL,
  crs = wtsGIS::get_crs(4326),
  nSZpG = 2,
  nPlotCols = 2
)
}
\arguments{
\item{sf_dfr}{- \pkg{sf} dataframe with geometries to plot}

\item{colours}{- vector of colours to use to distinguish start zones in an end location map}

\item{bmls}{- list of \pkg{ggplot2} basemap layers}

\item{crs}{- \pkg{sf} coordinate reference system (crs) for maps}

\item{nSZpG}{- number of start zones to include in a group for end location maps}

\item{nPlotCols}{- number of columns for end location maps}

\item{startZones}{- vector of starting zone id's (integers) for sf_zones}

\item{endZones}{- vector of ending zone id's (integers) for sf_zones}

\item{plotStartLocs}{- flag to plot start locations}

\item{plotEndLocs}{- flag to plot end locations}
}
\value{
a ggplot2 object
}
\description{
Function to create a ggplot2 object of maps from a \pkg{sf} dataframe.
}
\details{
Creates \pkg{ggplot2}-style maps of the starting and
ending locations of individuals. End zone locations can be colored according
to their start zone.
}
