% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_AssignZones.R
\name{indivsInfo_AssignZones}
\alias{indivsInfo_AssignZones}
\title{Assign zones to individuals based on location(s)}
\usage{
indivsInfo_AssignZones(
  sf_dfr,
  sf_zones,
  geomNames = c("startGeom", "endGeom"),
  zoneNames = c("startZone", "endZone"),
  keepFirst = TRUE
)
}
\arguments{
\item{sf_dfr}{- \pkg{sf} dataframe with point geometries}

\item{sf_zones}{- \pkg{sf} dataframe with zone information (polygons)}

\item{geomNames}{- vector of names of geometry column(s) to use for assigning zone(s)}

\item{zoneNames}{- vector of names of zones to be associated with each geomName}

\item{keepFirst}{- flag to keep first zone in which a point geometry falls (see Details)}
}
\value{
\code{sf_dfr} with additional column(s) indicating the zone(s) in which each geometry falls
}
\description{
Function to assign zones to individuals based on location(s).
}
\details{
Ideally, \code{sf_zones} should consist of non-overlapping polygons so that
only one zone can be assigned to each geometry. If a point geometry is in an area of overlap
between multiple zones, a record would normally be included in the result for each zone
in which the point falls. If \code{keepFirst} is TRUE, the record for the first zone
in which the point falls is kept. This is also the case if zones are assigned to multiple
\code{geomNames}. Otherwise, records are added to the result for all zones in which a point falls.
}
