% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ClassifyBySuccess.R
\name{indivsInfo_ClassifyBySuccess}
\alias{indivsInfo_ClassifyBySuccess}
\title{Classify an indivsInfo dataframe by individual success}
\usage{
indivsInfo_ClassifyBySuccess(dfrIndivs, dfrSuccess)
}
\arguments{
\item{dfrIndivs}{- dataframe with indivsInfo (e.g., from call to \code{\link{indivsInfo_ReorderResults}})}

\item{dfrSuccess}{- dataframe from \code{\link{indivsInfo_ExtractOriginalsBySuccess}}}
}
\value{
Dataframe identical to \code{dfrIndivs}, but with a column "successful" indicating
whether or not an individual was successful.
}
\description{
Function to classify an indivsInfo dataframe by individual success.
}
\details{
None.
}
