% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ExtractEnd.R
\name{indivsInfo_ExtractEnd}
\alias{indivsInfo_ExtractEnd}
\title{Extract info by individual at end of life or model run}
\usage{
indivsInfo_ExtractEnd(sf_EndByStage)
}
\arguments{
\item{sf_EndByStage}{- \pkg{sf} dataframe returned by \code{\link{indivsInfo_ExtractEndByStage}}}
}
\value{
\pkg{sf} dataframe with columns:
\itemize{
 \item{startTime}
 \item{origID - original (starting) individual ID}
 \item{endTime - ending time}
 \item{endID - ending individual ID}
 \item{endLHS - ending life stage}
 \item{endGridCell - ID of ending ROMS grid cell}
 \item{endLon - ending longitude}
 \item{endLat - ending latitude}
 \item{endDepth - ending depth (m)}
 \item{endBathym - bathymetric depth (m) at ending location}
 \item{endAge - ending age(d)}
 \item{endNum - ending number}
 \item{successful - flag indicating "success" (TRUE) or failure (FALSE) (e.g., settlement)}
 \item{endGeom - ending 2d location as SF_POINT}
}
}
\description{
Function to extract info by individual at end of life or model run.
}
\details{
The input \pkg{sf} dataframe should be the output of
\code{\link{indivsInfo_ExtractEndByStage}}.
For each unique individual, the end of the model run or death is identified by the record
where \code{age==max(age)}. "Unique" individuals are determined by eliminating original
individuals which resulted in more than one final individual (e.g., when metamorphosis
results in more than one individual).
}
