% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ExtractMaxAge.R
\name{indivsInfo_ExtractMaxAge}
\alias{indivsInfo_ExtractMaxAge}
\title{Extract age for each individual at end of life or model run}
\usage{
indivsInfo_ExtractMaxAge(sf_EndByStage)
}
\arguments{
\item{sf_EndByStage}{- \pkg{sf} dataframe returned by \code{\link{indivsInfo_ExtractEndByStage}} or \code{\link{indivsInfo_ExtractStartEndByStage}}}
}
\value{
\pkg{tibble} dataframe with columns:
\itemize{
 \item{startTime}
 \item{origID - original (starting) individual ID}
 \item{endID - individual ID at end}
 \item{maxAge - ending age(d)}
}
}
\description{
Function to extract age for each individual at end of life or model run.
}
\details{
The input \pkg{sf} dataframe should be the output of
\code{\link{indivsInfo_ExtractEndByStage}} or \code{\link{indivsInfo_ExtractStartEndByStage}}.

The end of the model run or death is basically identified by the record
where \code{age==max(age)} for each origID, endID pair. However, a straightforward
application of this approach would "double count"
individuals which changed id's during the course of their lifetime, because the "original"
individual (with id = origID) appears to have died when it metamorphoses (origID stays
the same, but id changes to a new unique value). Consequently, individuals which have
origID=endID are removed if there are any other individuals with the same origID but
different endIDs.
}
