% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ExtractStartEnd.R
\name{indivsInfo_ExtractStartEnd}
\alias{indivsInfo_ExtractStartEnd}
\title{Extract info by individual at start and end of life or model run}
\usage{
indivsInfo_ExtractStartEnd(
  sf_sebs = NULL,
  sf_start = NULL,
  sf_end = NULL,
  lst_indivs = NULL,
  startLHS = NULL,
  addVars = "",
  checkCalcs = FALSE
)
}
\arguments{
\item{sf_sebs}{- (optional) \pkg{sf} dataframe returned by \code{\link{indivsInfo_ExtractStartEndByStage}}}

\item{sf_start}{- (optional) \pkg{sf} dataframe returned by \code{\link{indivsInfo_ExtractStart}}}

\item{sf_end}{- (optional) \pkg{sf} dataframe returned by \code{\link{indivsInfo_ExtractEnd}}}

\item{lst_indivs}{- (req'd if sf_ebs and sf_start or sf_end is NULL) list of \pkg{sf} dataframes by life stage returned by \code{\link{indivsInfo_ReorderResults}}}

\item{startLHS}{- (req'd if sf_start is NULL) life stage in which all individuals start}

\item{addVars}{- (req'd if lst_indivs is given) character vector with names of additional (non-default) variables to extract}

\item{checkCalcs}{- flag (T/F) to check endGeom is correctly assigned (for debugging)}
}
\value{
\pkg{sf} dataframe with columns:
\itemize{
 \item{startTime}
 \item{origID - original (starting) individual ID}
 \item{startLHS - starting life stage}
 \item{startGridCell - ID of starting ROMS grid cell}
 \item{startLon - starting longitude}
 \item{startLat - starting latitude}
 \item{startDepth - starting depth (m)}
 \item{startBathym - bathymetric depth (m) at starting location}
 \item{startAge - starting depth}
 \item{startNum - starting depth}
 \item{start..AddVars - starting values for additional (non-default) variables}
 \item{startGeom - starting 2d location as SF_POINT}
 \item{endTime - ending time}
 \item{endID - ending individual ID}
 \item{endLHS - ending life stage}
 \item{endGridCell - ID of ending ROMS grid cell}
 \item{endLon - ending longitude}
 \item{endLat - ending latitude}
 \item{endDepth - ending depth (m)}
 \item{endBathym - bathymetric depth (m) at ending location}
 \item{endAge - ending age(d)}
 \item{endNum - ending number}
 \item{end..AddVars - ending values for additional (non-default) variables}
 \item{successful - flag indicating "success" (TRUE) or failure (FALSE) (e.g., settlement)}
 \item{endGeom - ending 2d location as SF_POINT}
}
}
\description{
Function to extract info by individual at start and end of life or model run.
}
\details{
If provided, the input \pkg{sf} dataframe \code{sf_sebs} should be the
output of \code{\link{indivsInfo_ExtractStartEndByStage}}. If \code{sf_sebs} is given
as an input, \code{addVars} is ignored.

If \code{sf_sebs} is NULL, the input \pkg{sf} dataframes \code{sf_start} and \code{sf_end}
can be given and should be the output of
\code{\link{indivsInfo_ExtractStart}} and  \code{\link{indivsInfo_ExtractEnd}}, respectively.


If either of these are NULL, \code{lst_indivs} must be given. This should be the list of
\pkg{sf} dataframes by life stage returned by \code{\link{indivsInfo_ReorderResults}}.


If \code{sf_start} is NULL, \code{startLHS} must be given.
If \code{sf_start} is NULL, it is calculated internally using
\code{sf_start = indivsInfo_ExtractStart(lst_indivs[[startLHS]],addVars=addVars)}.

If \code{sf_end} is NULL, it is calculated internally using
\code{sf_ebs = indivsInfo_ExtractEndByStage(lst_indivs,addVars=addVars)} followed by
\code{sf_end = indivsInfo_ExtractEnd(sf_ebs)}.
}
