% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ExtractStartEndByStage.R
\name{indivsInfo_ExtractStartEndByStage}
\alias{indivsInfo_ExtractStartEndByStage}
\title{Extract info by individual at start and end of life or model run}
\usage{
indivsInfo_ExtractStartEndByStage(
  sf_start = NULL,
  sf_ebs = NULL,
  lst_indivs = NULL,
  startLHS = NULL,
  addVars = "",
  checkCalcs = FALSE
)
}
\arguments{
\item{sf_start}{- (optional) \pkg{sf} dataframe returned by \code{\link{indivsInfo_ExtractStart}}}

\item{sf_ebs}{- (optional) \pkg{sf} dataframe returned by \code{\link{indivsInfo_ExtractEndByStage}}}

\item{lst_indivs}{- (req'd if sf_start or sf_end is NULL) list of \pkg{sf} dataframes by life stage returned by \code{\link{indivsInfo_ReorderResults}}}

\item{startLHS}{- (req'd if sf_start is NULL) life stage in which all individuals start}

\item{addVars}{- (req'd if lst_indivs is given) character vector with names of additional (non-default) variables to extract}

\item{checkCalcs}{- flag (T/F) to check endGeom is correctly assigned (for debugging)}
}
\value{
\pkg{sf} dataframe with columns:
\itemize{
 \item{startTime}
 \item{origID - original (starting) individual ID}
 \item{startLHS - starting life stage}
 \item{startGridCell - ID of starting ROMS grid cell}
 \item{startLon - starting longitude}
 \item{startLat - starting latitude}
 \item{startDepth - starting depth (m)}
 \item{startBathym - bathymetric depth (m) at starting location}
 \item{startAge - starting depth}
 \item{startNum - starting depth}
 \item{start..AddVars - starting values for additional (non-default) variables}
 \item{startGeom - starting 2d location as SF_POINT}
 \item{endTime - ending time}
 \item{endID - ending individual ID}
 \item{endLHS - ending life stage}
 \item{endGridCell - ID of ending ROMS grid cell}
 \item{endLon - ending longitude}
 \item{endLat - ending latitude}
 \item{endDepth - ending depth (m)}
 \item{endBathym - bathymetric depth (m) at ending location}
 \item{endAge - ending age(d)}
 \item{endNum - ending number}
 \item{end..AddVars - ending values for additional (non-default) variables}
 \item{successful - flag indicating "success" (TRUE) or failure (FALSE) (e.g., settlement)}
 \item{endGeom - ending 2d location as SF_POINT}
}
}
\description{
Function to extract info by individual at start and end of life or model run.
}
\details{
The input \pkg{sf} dataframes \code{sf_start} and \code{sf_end} should be
the output of \code{\link{indivsInfo_ExtractStart}} and  \code{\link{indivsInfo_ExtractEnd}},
respectively. If both are given as an inputs, \code{addVars} is ignored.

If either are NULL, \code{lst_indivs} must be given. If \code{sf_start} is NULL,
\code{startLHS} must be given. If \code{sf_start} is NULL, it is calculated internally
using \code{sf_start = indivsInfo_ExtractStart(lst_indivs[[startLHS]],addVars=addVars)}.

If \code{sf_end}
is NULL, it is calculated internally using \code{sf_ebs = indivsInfo_ExtractEndByStage(lst_indivs,addVars=addVars)}
followed by \code{sf_end = indivsInfo_ExtractEnd(sf_ebs)}.
}
