% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ExtractTrajectories.R
\name{indivsInfo_ExtractTrajectories}
\alias{indivsInfo_ExtractTrajectories}
\title{Extract trajectories from DisMELS output as a list of \pkg{sf} dataframes with linestring geometries}
\usage{
indivsInfo_ExtractTrajectories(dfrs, crs = wtsGIS::get_crs("WGS84"))
}
\arguments{
\item{dfrs}{- list of dataframes, by typeName, with DisMELS IBM results}

\item{crs}{- coordinate reference system: \pkg{sf} crs object, EPSG code, or character with proj4string}
}
\value{
a list of \pkg{sf} dataset objects, each with a column ("geom") of class sfc_LINESTRING
}
\description{
Function to extract trajectories from DisMELS output as a list of \pkg{sf} dataframes with linestring geometries.
}
\details{
Requires packages \code{sf}, \code{wtsGIS}. For each \pkg{sf} dataframe, the linestring geometry
is in column "geom". Other columns include id, startTime, age, ageInStage, and successful.
}
