% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ReorderResults.R
\name{indivsInfo_ReorderResults}
\alias{indivsInfo_ReorderResults}
\title{Reorder a set of IBM results by life stage, individual and age}
\usage{
indivsInfo_ReorderResults(
  resFolder,
  resFilePrefix,
  lifeStageInfo,
  startTime = NULL
)
}
\arguments{
\item{resFolder}{- path to folder with results}

\item{resFilePrefix}{- prefix for file names (default = "Results")}

\item{lifeStageInfo}{- life stage info (list) object}

\item{startTime}{- optional value to replace startTimes in model run (default=NULL)}
}
\value{
list of data frames by life stage type,
with each dataframe ordered by start time, id, and time.
}
\description{
Function to reorder a set of IBM results by life stage, individual and age.
}
\details{
Requires package \pkg{readr} to read csv files.
}
\note{
Times in the csv files are interpreted by readr::read_csv as UTC, as well
as in the reordered dataframes.

Assigning a value to the input \code{startTime} replaces all values of startTime
from the model run with the input value in the resulting output. As such, \code{startTime}
should only be given when, for example, the individuals in a DisMELS model run were initially released
at the same time, but the simulation was restarted with all individuals at the start of a
subsequent life stage (resulting in different startTimes for individuals in the DisMELS results for the
restarted simulation). Replacing startTime for results from the restarted simulation output yields
simulation results equivalent to a continuous model run (i.e., without the restart).
}
