% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap_ValuesByGridCell.R
\name{plotMap_ValuesByGridCell}
\alias{plotMap_ValuesByGridCell}
\title{Create a map of values by grid cell}
\usage{
plotMap_ValuesByGridCell(dfr, vals_col, roms_grid, basemap,
  title = "values", alpha = 0.9)
}
\arguments{
\item{dfr}{- dataframe with values to plot by gridCellID}

\item{roms_grid}{- shapefile name or layer (sf dataset) with grid polygons by ID}

\item{basemap}{- basemap to use}

\item{title}{- map title}

\item{alpha}{- transparency to apply to values layer}

\item{val_col}{- column name of values to plot}
}
\value{
a tmap map object of the values plotted by grid cell
}
\description{
Function to create a map of values by grid cell.
}
\details{
Uses packages \code{wtsGIS}, \code{tmap}. Note that the color scale used for the values
is determined by the \code{aes.palette} argument to \code{createBasemap} when the basemap is created.
}
