% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderAndClassifyResults.R
\name{reorderAndClassifyResults}
\alias{reorderAndClassifyResults}
\title{Reorder a set of IBM results by life stage and individual}
\usage{
reorderAndClassifyResults(
  resFolder,
  resFilePrefix,
  lifeStageInfo,
  dfrSuccess = NULL,
  crs = wtsGIS::get_crs(4326),
  wrapDateline = TRUE,
  zipFiles = FALSE
)
}
\arguments{
\item{resFolder}{- path to folder with results}

\item{resFilePrefix}{- prefix for file names (default = "Results")}

\item{lifeStageInfo}{- lhsInfo (list) object with life stage info}

\item{dfrSuccess}{- dataframe with individuals classified by success}

\item{crs}{- coordinate reference system (object that can be converted to an \code{sf::crs} object; default is WGS84)}

\item{wrapDateline}{- flag (T/F) to use 0 to 360 rather than -180 to 180 range for longitudes}

\item{zipFiles}{- flag to zip DisMELS csv files}
}
\value{
list of \pkg{sf} dataframes by life stage type,
with each dataframe ordered by start time, origID, and time. An
additional column, 'success', may be present and indicate ultimately
successful (TRUE) or unsuccessful (FALSE) individuals. The
'geom' column contains the point location of each observation.
}
\description{
Function to reorder a set of IBM results by life stage and individual.
}
\details{
Requires package \pkg{readr} to read csv files.
}
\note{
If locations are in projected coordinates, then \code{wrapDateline}
should be set to FALSE (the default is TRUE).
}
