% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderResultsByStageAndIndiv.R
\name{reorderResultsByStageAndIndiv}
\alias{reorderResultsByStageAndIndiv}
\title{Reorder a set of IBM results by life stage and individual}
\usage{
reorderResultsByStageAndIndiv(resFolder, resFilePrefix, lifeStageInfo)
}
\arguments{
\item{resFolder}{- path to folder with results}

\item{resFilePrefix}{- prefix for file names (default = "Results")}

\item{lifeStageInfo}{- lhsInfo (list) object with life stage info}
}
\value{
list of data frames by life stage type,
with each dataframe ordered by start time, id, and time.
}
\description{
Function to reorder a set of IBM results by life stage and individual.
}
\details{
Requires package \code{readr} to read csv files.
}
