% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_CreateLayer_ValuesByGridCell.R
\name{tmap_CreateLayer_ValuesByGridCell}
\alias{tmap_CreateLayer_ValuesByGridCell}
\title{Create a \pkg{tmap}-style map layer with values by grid cell}
\usage{
tmap_CreateLayer_ValuesByGridCell(
  dfr,
  vals_col,
  basemap = NULL,
  title = "values",
  alpha = 0.9,
  plotMap = FALSE
)
}
\arguments{
\item{dfr}{- \code{sf} dataframe with values to create map layer by gridCellID}

\item{vals_col}{- column name of values to plot}

\item{basemap}{- \pkg{tmap}-style basemap to use (default is NULL, which uses the basemap for the CGOA in lat/lon)}

\item{title}{- map title (default is "values")}

\item{alpha}{- transparency to apply to values layer (default is 0.9)}

\item{plotMap}{- flag to combine layer with basemap and plot (default is FALSE)}
}
\value{
a \pkg{tmap}-style layer object of the values by grid cell.
}
\description{
Function to create a \pkg{tmap}-style map layer with values by grid cell.
}
\details{
Requires package \pkg{tmap}. If all values are 0, then NULL is
returned for the map layer object. Note that the color scale used for the values
is determined by the \code{aes.palette} argument to \code{createBasemap} when the basemap is created.
}
