% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{toCamelCase}
\alias{toCamelCase}
\title{Convert strings to camel case}
\usage{
toCamelCase(x)
}
\arguments{
\item{x}{- character vector to convert}
}
\value{
character vector
}
\description{
Function to convert strings to camel case
}
