% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateBBox.R
\name{updateBBox}
\alias{updateBBox}
\title{Update a bbox (bounding box)}
\usage{
updateBBox(bbx1 = NULL, bbx2 = NULL)
}
\arguments{
\item{bbx1}{- original bbox}

\item{bbx2}{- updating bbox}
}
\value{
resulting bbox
}
\description{
Function to update a bbox (bounding box)
}
\details{
Returns a bbox covering the union of the individual bboxes.
}
