% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPointGeometries.R
\name{addPointGeometries}
\alias{addPointGeometries}
\title{Add point geometries to a list dataframes from DisMELS output}
\usage{
addPointGeometries(dfrs)
}
\arguments{
\item{dfrs}{- list of dataframes, by typeName, with DisMELS IBM results}

\item{strCRS}{- coordinate reference system: EPSG code or character with proj4string}
}
\value{
a list of sf (dataset) objects, by typeName, each with a column ("geom") of class sfc_POINT
}
\description{
Function to add point geometries to a list dataframes from DisMELS output
}
\details{
Uses packages \code{sf}, \code{wtsGIS}. For each sf dataset, the point geometries
are in column "geom". Other columns are copied from the associated input dataframes.
}
