% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byGridCell_CalcStats.R
\name{byGridCell_CalcStats}
\alias{byGridCell_CalcStats}
\title{Calculate statistiics for results by grid cell}
\usage{
byGridCell_CalcStats(dfrs)
}
\arguments{
\item{dfrs}{- list of dataframes with results by grid cell}
}
\value{
dataframe with statistics
}
\description{
Function to calculate statistiics for results by grid cell.
}
\details{
TODO: This function does nothing, Finish implementing it!
}
