% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byGridCell_CreateSFDataset.R
\name{byGridCell_CreateSFDataset}
\alias{byGridCell_CreateSFDataset}
\title{Create a map layer with values by grid cell}
\usage{
byGridCell_CreateSFDataset(dfr, roms_grid)
}
\arguments{
\item{dfr}{- dataframe with values to create map layer by gridCellID}

\item{roms_grid}{- shapefile name or layer (sf dataset) with grid polygons by ID}
}
\value{
an \code{sf} (simple features) dataframe with the values by grid cell.
}
\description{
Function to create a map layer with values by grid cell.
}
\details{
Uses package \code{wtsGIS}. Cells in the \code{roms_grid} which are not matched
in the dataframe \code{dfr} will have NAs in the corresponding columns from
\code{dfr}.
}
