% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyIndivsBySuccess.R
\name{classifyIndivsBySuccess}
\alias{classifyIndivsBySuccess}
\title{Classify individuals by success/failure}
\usage{
classifyIndivsBySuccess(dfrs, typeName)
}
\arguments{
\item{dfrs}{- list of dataframes by typeName}

\item{typeName}{- life stage typeName indicating success}
}
\value{
the list of dataframes by typeName, with added column "succcess"
(TRUE/FALSE) indicating whether individual is considered a success.
}
\description{
Function to classify individuals by success/failure.
}
\details{

}
