% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMapLayer_ValuesByGridCell.R
\name{createMapLayer_ValuesByGridCell}
\alias{createMapLayer_ValuesByGridCell}
\title{Create a map layer with values by grid cell}
\usage{
createMapLayer_ValuesByGridCell(
  dfr,
  vals_col,
  roms_grid,
  basemap = NULL,
  title = "values",
  alpha = 0.9,
  plotMap = FALSE
)
}
\arguments{
\item{dfr}{- dataframe with values to create map layer by gridCellID}

\item{roms_grid}{- shapefile name or layer (sf dataset) with grid polygons by ID}

\item{basemap}{- basemap to use}

\item{title}{- map title}

\item{alpha}{- transparency to apply to values layer}

\item{plotMap}{- flag to combine layer with basemap and plot}

\item{val_col}{- column name of values to plot}
}
\value{
a tmap layer object of the values by grid cell, or NULL if all values are 0.
}
\description{
Function to create a map layer with values by grid cell.
}
\details{
Uses packages \code{wtsGIS}, \code{tmap}. Note that the color scale used for the values
is determined by the \code{aes.palette} argument to \code{createBasemap} when the basemap is created.
}
