% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTrajectories.R
\name{extractTrajectories}
\alias{extractTrajectories}
\title{Extract trajectories from DisMELS output as a list of sf (simple features) datasets (dataframes) with linestring geometries}
\usage{
extractTrajectories(dfrs, strCRS = wtsGIS::getCRS("WGS84"))
}
\arguments{
\item{dfrs}{- list of dataframes, by typeName, with DisMELS IBM results}

\item{strCRS}{- coordinate reference system: EPSG code or character with proj4string}
}
\value{
a list of sf dataset objects, each with a column ("geom") of class sfc_LINESTRING
}
\description{
Function to extract trajectories from DisMELS output as a list of sf (simple features) datasets (dataframes) with linestring geometries.
}
\details{
Requires packages \code{sf}, \code{wtsGIS}. For each sf dataset, the linestring geometry
is in column "geom". Other columns include id, startTime, age, ageInStage, and successful.
}
