% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_extractIndivs.R
\name{extractTrajectories}
\alias{extractTrajectories}
\title{Extract trajectories of (selected) individuals from a DisMELS model run}
\usage{
extractTrajectories(results = NULL, indivIDs = NULL,
  lsTypesInfo = NULL, lsTypes = NULL, returnList = TRUE,
  writeOutput = TRUE, outDir = "./", outBaseCSV = "allIndivs",
  verbose = FALSE)
}
\arguments{
\item{results}{- DisMELS results file (or dataframe)}

\item{indivIDs}{- vector of id's for individuals to extract (NULL = all)}

\item{lsTypesInfo}{- life stage types info list for IBM}

\item{lsTypes}{- vector of names of life stage types to extract (NULL = all)}

\item{returnList}{- flag (T/F) indicating whether a list should be returned (otherwise NULL--helps with memory)}

\item{writeOutput}{- flag (T/F) indicating whether output files should be written}

\item{outDir}{- folder for output}

\item{outBaseCSV}{- base name for output csv files}

\item{verbose}{- flag (T/F) to print debugging info}
}
\value{
list of dataframes with individual results by lhs type name if returnList=TRUE, otherwise NULL.
}
\description{
Function to extract trajectories of (selected) individuals from a DisMELS model run.
}
\details{
If results is NULL, the user will be prompted to select the
corresponding file using a file chooser dialog.
Uses functions \code{wtsUtilities::getCSV} and \code{sqldf::sqldf}.
}
