% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMap_Connectivity.R
\name{ggMap_Connectivity}
\alias{ggMap_Connectivity}
\title{Create a \pkg{ggplot2} map object of a connectivity matrix}
\usage{
ggMap_Connectivity(
  tbl_conn,
  tbl_web,
  bmls = NULL,
  valCol = "value",
  facetCol = "startTime",
  label = "value",
  val_min = 0,
  bin_size = NULL,
  bin_max = NULL,
  max_size = 3,
  colour_scale = ggplot2::scale_colour_viridis_c(option = "plasma"),
  arrow = grid::arrow(20, grid::unit(0.15, "inches")),
  plotRetentionOnly = FALSE,
  plotExportedOnly = FALSE
)
}
\arguments{
\item{tbl_conn}{-  tibble with connectivity values to plot}

\item{tbl_web}{- tibble representing the connectivity web}

\item{bmls}{- list of base map layers {\pkg{ggplot2} layer objects}}

\item{valCol}{- column name for values to plot}

\item{facetCol}{- column name for faceting variable}

\item{label}{- label for legends}

\item{val_min}{- minimum value to include (0 is never included)}

\item{bin_size}{- bin size (no binning if NULL)}

\item{bin_max}{- max bin value, if binning applied}

\item{max_size}{- max size for connectivity arrows}

\item{colour_scale}{- \pkg{ggplot2} colour scale for connectivity values}

\item{arrow}{- pass result from call to \code{\link[grid]{arrow}}}

\item{plotRetentionOnly}{- flag (T/F) to plot only retained proportions}

\item{plotExportedOnly}{- flag (T/F) to plot only non-retained (exported) proportions}
}
\value{
\pkg{ggplot2} map (plot) object
}
\description{
Function to create a \pkg{ggplot2} map object of a connectivity matrix.
}
\details{
Creates a \pkg{ggplot2} map object representing a connectivity matrix
as directed arrows from start zone locations to end zone locations.
}
