% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMap_Connectivity.R
\name{ggMap_Connectivity}
\alias{ggMap_Connectivity}
\title{Create a \pkg{ggplot2} map object of a connectivity matrix}
\usage{
ggMap_Connectivity(
  tbl_conn,
  tbl_web,
  bmls = NULL,
  maxVal = NA,
  valCol = "value",
  facetGrid = NULL,
  facetWrap = NULL,
  facetByRow = TRUE,
  nrow = NULL,
  ncol = NULL,
  label = "value",
  val_min = 0,
  bin_size = NULL,
  bin_max = NULL,
  max_size = 3,
  linewidth = NA,
  alpha = NA,
  reverse = TRUE,
  scale_colour = ggplot2::scale_colour_viridis_c(option = "plasma", limits = c(0,
    maxVal), oob = scales::squish),
  scale_size = ggplot2::scale_size_area(max_size = max_size, oob = scales::squish),
  scale_alpha = ggplot2::scale_alpha_continuous(oob = scales::squish),
  arrow = grid::arrow(20, grid::unit(0.15, "inches")),
  noAxisLabels = TRUE,
  plotRetentionOnly = FALSE,
  plotExportedOnly = FALSE
)
}
\arguments{
\item{tbl_conn}{-  tibble with connectivity values to plot}

\item{tbl_web}{- tibble representing the connectivity web}

\item{bmls}{- list of base map layers {\pkg{ggplot2} layer objects}}

\item{maxVal}{- max value for scales (higher values will be set to maxVal)}

\item{valCol}{- column name for values to plot}

\item{facetGrid}{- formula for faceting variables using \code{ggplot2::facet_grid}}

\item{facetWrap}{- column name(s) for faceting variable(s) using \code{ggplot2::facet_wrap}}

\item{facetByRow}{- flag (T/F) to arrange facets by row}

\item{nrow}{- number of rows to arrange facets in (if facetByRow is TRUE)}

\item{ncol}{- number of columns to arrange facets in (if facetByRow is FALSE)}

\item{label}{- label for legends}

\item{val_min}{- minimum value to include (0 is never included)}

\item{bin_size}{- bin size (no binning if NULL)}

\item{bin_max}{- max bin value, if binning applied}

\item{max_size}{- max size for connectivity arrows}

\item{linewidth}{- line width for fixed-width arrows}

\item{alpha}{- transparency value for fixed-alpha symbols and arrows}

\item{reverse}{- plot connections in reverse order (largest on bottom)}

\item{scale_colour}{- \pkg{ggplot2} colour scale for connectivity values (default = ggplot2::scale_colour_viridis_c(option="plasma",limits=c(0,maxVal)))}

\item{scale_size}{- \pkg{ggplot2} size scale for connectivity values (default = scale_size_area(max_size=max_size,oob=scales::squish))}

\item{scale_alpha}{- \pkg{ggplot2} transparency scale for connectivity values (default = scale_alpha_continuous(oob=scales::squish))}

\item{arrow}{- pass result from call to \code{\link[grid]{arrow}}}

\item{noAxisLabels}{- flag (T/F) to include axis labels}

\item{plotRetentionOnly}{- flag (T/F) to plot only retained proportions}

\item{plotExportedOnly}{- flag (T/F) to plot only non-retained (exported) proportions}
}
\value{
\pkg{ggplot2} map (plot) object
}
\description{
Function to create a \pkg{ggplot2} map object of a connectivity matrix.
}
\details{
Creates a \pkg{ggplot2} map object representing a connectivity matrix
as directed arrows from start zone locations to end zone locations.
}
