% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ClassifyBySuccess.R
\name{indivsInfo_ClassifyBySuccess}
\alias{indivsInfo_ClassifyBySuccess}
\title{Extract IDs for successful individuals}
\usage{
indivsInfo_ClassifyBySuccess(dfrStart, dfrEnd, typeNames)
}
\arguments{
\item{dfrStart}{- a dataframe or tibble that includes individuals in the initial life stage}

\item{dfrEnd}{- a dataframe or tibble that includes individuals in the "successful" life stage}

\item{typeNames}{- vector of names of life stages defining success}
}
\value{
a dataframe or tibble with origID, startTime, and successful for all individuals.
}
\description{
Function to extract IDs for successful individuals.
}
\details{
The
}
