% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ExtractMaxAge.R
\name{indivsInfo_ExtractMaxAge}
\alias{indivsInfo_ExtractMaxAge}
\title{Extract age for each individual at end of life or model run}
\usage{
indivsInfo_ExtractMaxAge(sf_EndByStage)
}
\arguments{
\item{sf_EndByStage}{- \pkg{sf} dataframe returned by \code{\link{indivsInfo_ExtractEndByStage}} or \code{\link{indivsInfo_ExtractStartEndByStage}}}
}
\value{
\pkg{tibble} dataframe with columns:
\itemize{
 \item{startTime}
 \item{origID - original (starting) individual ID}
 \item{endID - individual ID at end}
 \item{maxAge - ending age(d)}
}
}
\description{
Function to extract age for each individual at end of life or model run.
}
\details{
The input \pkg{sf} dataframe should be the output of
\code{\link{indivsInfo_ExtractEndByStage}} or \code{\link{indivsInfo_ExtractStartEndByStage}}.
For each unique individual, the end of the model run or death is identified by the record
where \code{age==max(age)}. "Unique" individuals are determined by eliminating original
individuals which resulted in more than one final individual (e.g., when metamorphosis
results in more than one individual).
}
