% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ExtractOriginalsBySuccess.R
\name{indivsInfo_ExtractOriginalsBySuccess}
\alias{indivsInfo_ExtractOriginalsBySuccess}
\title{Extract a dataframe of original IDs by success}
\usage{
indivsInfo_ExtractOriginalsBySuccess(dfrStart, dfrEnd)
}
\arguments{
\item{dfrStart}{- dataframe with startTime and origID of unique initial individuals}

\item{dfrEnd}{- dataframe with startTime, orgiID, and id of unique individuals in "successful" life stage(s)}
}
\value{
a dataframe with origID and startTime for successful individuals.
}
\description{
Function to extract a dataframe of original IDs by success.
}
\details{
The "index" consists of a dataframe with columns
\itemize{
  \item{startTime}
  \item{origID}
  \item{successful - TRUE/FALSE indicating whether or not a "succssful" life stage was eventually reached}
}
and a row for each "original" individual.
}
