% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ExtractStart.R
\name{indivsInfo_ExtractStart}
\alias{indivsInfo_ExtractStart}
\title{Extract info on individuals at start of initial life stage}
\usage{
indivsInfo_ExtractStart(sf_indivs, addVars = "")
}
\arguments{
\item{sf_indivs}{- \pkg{sf} dataframe for the initial life stage from call to \code{\link{indivsInfo_ReorderResults}}}

\item{addVars}{- character vector with names of additional (non-default) variables to extract}
}
\value{
\pkg{sf} dataframe with columns:
\itemize{
 \item{startTime}
 \item{origID - original (starting) individual ID}
 \item{startLHS - starting life stage}
 \item{startGridCell - ID of starting ROMS grid cell}
 \item{startLon - starting longitude}
 \item{startLat - starting latitude}
 \item{startDepth - starting depth (m)}
 \item{startBathym - bathymetric depth (m) at starting location}
 \item{startAge - starting depth}
 \item{startNum - starting depth}
 \item{startTemp - starting temperature}
 \item{start..AddVars - columns corresponding to additional variables}
 \item{successful - flag indicating "success" (TRUE) or failure (FALSE) (e.g., settlement)}
 \item{startGeom - starting 2d location as SF_POINT}
}
}
\description{
Function to extract info on individuals at start of initial life stage.
}
\details{
The input \pkg{sf} dataframe should be the one returned
by \code{\link{indivsInfo_ReorderResults}} for the initial life stage.
For each individual, the initial location is identified by the record
when \code{startTime == time}.
}
