% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indivsInfo_ReorderResults.R
\name{indivsInfo_ReorderResults}
\alias{indivsInfo_ReorderResults}
\title{Reorder a set of IBM results by life stage, individual and age}
\usage{
indivsInfo_ReorderResults(resFolder, resFilePrefix, lifeStageInfo)
}
\arguments{
\item{resFolder}{- path to folder with results}

\item{resFilePrefix}{- prefix for file names (default = "Results")}

\item{lifeStageInfo}{- life stage info (list) object}
}
\value{
list of data frames by life stage type,
with each dataframe ordered by start time, id, and time.
}
\description{
Function to reorder a set of IBM results by life stage, individual and age.
}
\details{
Requires package \pkg{readr} to read csv files.
}
\note{
Times in the csv files are interpreted by readr::read_csv as UTC, as well
as in the reordered dataframes.
}
