% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTracks.R
\name{parseTracks}
\alias{parseTracks}
\title{Create an sf (simple features) column with linestring geometries from DisMELS tracks}
\usage{
parseTracks(tracks, strCRS = wtsGIS::getCRS("WGS84"))
}
\arguments{
\item{tracks}{- track column from DisMELS model results file}

\item{strCRS}{- coordinate reference system: EPSG code or character with proj4string}
}
\value{
an object of class sfc_LINESTRING
}
\description{
Function to create an sf (simple features) column with linestring geometries from DisMELS tracks.
}
\details{
Requires packages \code{sf}, \code{wtsGIS}.
}
