% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConnectivityMatrix.R
\name{plotConnectivityMatrix}
\alias{plotConnectivityMatrix}
\title{Create a \pkg{ggplot2} connectivity matrix plot object}
\usage{
plotConnectivityMatrix(
  tbl_conn,
  connPolys,
  maxVal = NA,
  valCol = "value",
  facetGrid = NULL,
  facetWrap = NULL,
  facetByRow = TRUE,
  nrow = NULL,
  ncol = NULL,
  setZeroToNA = FALSE,
  xlab = "end zone",
  ylab = "start zone",
  fill_lab = "value",
  fill_scale = ggplot2::scale_fill_viridis_c(option = "plasma", limits = c(0, maxVal)),
  legend.position = "right"
)
}
\arguments{
\item{tbl_conn}{-  tibble with connectivity values to plot}

\item{connPolys}{- tibble representing the connectivity grid}

\item{maxVal}{- max value for fill scale}

\item{valCol}{- column name for values to plot}

\item{facetGrid}{- formula for faceting variables using \code{ggplot2::facet_grid}}

\item{facetWrap}{- column name(s) for faceting variable(s) using \code{ggplot2::facet_wrap}}

\item{facetByRow}{- flag (T/F) to arrange wrapped facets by row}

\item{nrow}{- number of rows to arrange wrapped facets in (if facetByRow is TRUE)}

\item{ncol}{- number of columns to arrange wrapped facets in (if facetByRow is FALSE)}

\item{setZeroToNA}{- flag (T/F) to set 0 values to NA's}

\item{xlab}{- x axis label}

\item{ylab}{- x axis label}

\item{fill_lab}{- label for fill legend}

\item{fill_scale}{- \pkg{ggplot2} fill scale}

\item{legend.position}{- string indicating legend position in \code{\link[ggplot2]{theme}}}
}
\value{
\pkg{ggplot2} plot object
}
\description{
Function to create a \pkg{ggplot2} connectivity matrix plot object.
}
\details{
Creates a \pkg{ggplot2} plot object representing a connectivity matrix
with start zones on the y axis and end zones on the x axis. The startZone and endZone
columns of the connPolys dataframe should be factors, with levels corresponding to
the order in which zones should be displayed.
}
