% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderAndClassifyResults.R
\name{reorderAndClassifyResults}
\alias{reorderAndClassifyResults}
\title{Reorder a set of IBM results by life stage and individual}
\usage{
reorderAndClassifyResults(
  resFolder,
  resFilePrefix,
  lifeStageInfo,
  dfrSuccess = NULL
)
}
\arguments{
\item{resFolder}{- path to folder with results}

\item{resFilePrefix}{- prefix for file names (default = "Results")}

\item{lifeStageInfo}{- lhsInfo (list) object with life stage info}

\item{dfrSuccess}{- dataframe with individuals classified by success}
}
\value{
list of \pkg{sf} dataframes by life stage type,
with each dataframe ordered by start time, origID, and time. An
additional column, 'success', may be present and indicate ultimately
successful (TRUE) or unsuccessful (FALSE) individuals. The
'geom' column contains the point location of each observation.
}
\description{
Function to reorder a set of IBM results by life stage and individual.
}
\details{
Requires package \pkg{readr} to read csv files.
}
