% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipDisMELSResults.R
\name{zipDisMELSResults}
\alias{zipDisMELSResults}
\title{Reorder a set of IBM results by life stage and individual}
\usage{
zipDisMELSResults(resFolder, resFilePrefix, lifeStageInfo)
}
\arguments{
\item{resFolder}{- path to folder with results}

\item{resFilePrefix}{- prefix for file names (default = "Results")}

\item{lifeStageInfo}{- lhsInfo (list) object with life stage info}
}
\value{
none.
}
\description{
Function to reorder a set of IBM results by life stage and individual.
}
\details{
Identified DisMELS files are zipped. Originals are not deleted.
}
